⚙️ refactor(recipes): Standardize ore dictionary usage and file structure
Updates recipes to use standardized ore dictionary entries and reorganizes JSON files into logical subdirectories for improved maintainability.
