+++ src/main/resources/data/okprogressions/recipes/crafting/blocks/botanypot/botany_pot.json
