+++ src/main/resources/data/okprogressions/recipes/crafting/blocks/botanypot/hopper_botany_pot_2.json
