+++ src/main/resources/data/okprogressions/recipes/crafting/blocks/cobblegen/diamond_cobblegen.json
@@ -1,0 +1,20 @@
+{
+  "type": "minecraft:crafting_shaped",
+  "pattern": [
+    "iii",
+    "igi",
+    "iii"
+  ],
+  "key": {
+    "i": {
+      "ore": "#forge:gems/diamond"
+    },
+    "g": {
+      "item": "okprogressions:gold_cobblegen"
+    }
+  },
+  "result": {
+    "item": "okprogressions:diamond_cobblegen",
+    "amount": 1
+  }
+}
