+++ src/main/resources/data/okprogressions/recipes/crafting/blocks/compressed/charcoal_block.json
