+++ src/main/resources/data/okprogressions/recipes/crafting/blocks/compressed/flint_block.json
