+++ src/main/resources/data/okprogressions/recipes/crafting/blocks/growth/growth.json
@@ -7,7 +7,7 @@
   ],
   "key": {
     "i": {
-      "item": "minecraft:iron_ingot"
+      "ore": "#forge:ingots/iron"
     },
     "b": {
       "item": "okprogressions:bone_block"
