+++ src/main/resources/data/okprogressions/recipes/crafting/blocks/growth/growth_upgrade_two.json
