+++ src/main/resources/data/okprogressions/recipes/crafting/blocks/reinforced/reinforced_obsidian.json
