--- src/main/resources/data/okprogressions/recipes/crafting/blocks/stone_torch_charcoal.json
@@ -1,20 +1,0 @@
-{
-  "type": "minecraft:crafting_shaped",
-  "pattern": [
-    "c",
-    "s"
-  ],
-  "key": {
-    "c": {
-      "item": "minecraft:coal",
-      "meta": 1
-    },
-    "s": {
-      "item": "okprogressions:stone_stick"
-    }
-  },
-  "result": {
-    "item": "okprogressions:stone_torch",
-    "amount": 4
-  }
-}
