+++ src/main/resources/data/okprogressions/recipes/crafting/items/misc/charcoal_to_tiny_charcoal.json
