+++ src/main/resources/data/okprogressions/recipes/crafting/items/misc/coal_to_tiny_coal.json
