+++ src/main/resources/data/okprogressions/recipes/crafting/items/misc/ender_sack.json
@@ -7,11 +7,10 @@
   ],
   "key": {
     "s": {
-      "item": "minecraft:ender_chest"
+      "ore": "#forge:chests/ender"
     },
     "l": {
-      "item": "minecraft:leather",
-      "meta": 0
+      "item": "minecraft:leather"
     }
   },
   "result": {
