+++ src/main/resources/data/okprogressions/recipes/crafting/items/misc/stone_stick.json
@@ -6,7 +6,7 @@
   ],
   "key": {
     "c": {
-      "ore": "cobblestone"
+      "ore": "#forge:cobblestone/normal"
     }
   },
   "result": {
