+++ src/main/resources/data/okprogressions/recipes/crafting/items/misc/tiny_charcoal_to_charcoal.json
