+++ src/main/resources/data/okprogressions/recipes/crafting/items/misc/tiny_coal_to_coal.json
