--- src/main/resources/data/okprogressions/recipes/crafting/items/redstone_apple.json
@@ -1,20 +1,0 @@
-{
-  "type": "minecraft:crafting_shaped",
-  "pattern": [
-    "rrr",
-    "rar",
-    "rrr"
-  ],
-  "key": {
-    "r": {
-      "ore": "dustRedstone"
-    },
-    "a": {
-      "item": "minecraft:apple"
-    }
-  },
-  "result": {
-    "item": "okprogressions:redstone_apple",
-    "amount": 1
-  }
-}
