+++ src/main/resources/data/okprogressions/recipes/crafting/items/tool/wooden_paxel.json
@@ -16,7 +16,7 @@
       "item": "minecraft:wooden_pickaxe"
     },
     "s": {
-      "ore": "stickWood"
+      "ore": "#forge:rods/wooden"
     }
   },
   "result": {
